feature: redesigned the hotky and target UIs so you can now create new targets in the hotkey edit ui.
