+++ tests/svelte/HotkeysTab.test.ts
@@ -1,6 +1,6 @@
 import { describe, test, expect, vi, beforeEach } from "vitest";
 import { render, screen, fireEvent } from "@testing-library/svelte";
-import RoutingTab from "../../src/lib/Settings/RoutingTab.svelte";
+import HotkeysTab from "../../src/lib/Settings/HotkeysTab.svelte";
 import type { HotkeyBinding, OutputTarget } from "../../src/lib/Settings/routing-types";
 
 let mockBindings: HotkeyBinding[] = [];
@@ -19,7 +19,7 @@ vi.mock("@tauri-apps/api/core", () => ({
   }),
 }));
 
-describe("RoutingTab.svelte Conflict Detection", () => {
+describe("HotkeysTab.svelte Conflict Detection and Nested Modal", () => {
   beforeEach(() => {
     mockTargets = [
       {
@@ -36,11 +36,6 @@ describe("RoutingTab.svelte Conflict Detection", () => {
     ];
     mockBindings = [];
   });
-
-  async function switchToBindingsTab() {
-    const tabButton = await screen.findByRole("button", { name: /Hotkey Bindings/i });
-    await fireEvent.click(tabButton);
-  }
 
   test("does not show conflict warnings when there are no conflicts", async () => {
     mockBindings = [
@@ -68,8 +63,7 @@ describe("RoutingTab.svelte Conflict Detection", () => {
       },
     ];
 
-    render(RoutingTab);
-    await switchToBindingsTab();
+    render(HotkeysTab);
 
     // Conflict banner should NOT be present
     const banner = screen.queryByText(/Conflict detected/i);
@@ -106,8 +100,7 @@ describe("RoutingTab.svelte Conflict Detection", () => {
       },
     ];
 
-    const { container } = render(RoutingTab);
-    await switchToBindingsTab();
+    const { container } = render(HotkeysTab);
 
     // Conflict banner should be present
     const banner = await screen.findByText(/Conflict detected/i);
@@ -148,8 +141,7 @@ describe("RoutingTab.svelte Conflict Detection", () => {
       },
     ];
 
-    const { container } = render(RoutingTab);
-    await switchToBindingsTab();
+    const { container } = render(HotkeysTab);
 
     // Conflict banner should still be present because a conflict exists in the list
     const banner = await screen.findByText(/Conflict detected/i);
@@ -183,8 +175,7 @@ describe("RoutingTab.svelte Conflict Detection", () => {
       },
     ];
 
-    render(RoutingTab);
-    await switchToBindingsTab();
+    render(HotkeysTab);
 
     const badge = await screen.findByText(/Ollama LLM/i);
     expect(badge).not.toBeNull();
@@ -206,10 +197,114 @@ describe("RoutingTab.svelte Conflict Detection", () => {
       },
     ];
 
-    render(RoutingTab);
-    await switchToBindingsTab();
+    render(HotkeysTab);
 
     const badge = screen.queryByText(/Ollama LLM/i);
     expect(badge).toBeNull();
+  });
+
+  test("opens nested Target modal and cancels to revert select value", async () => {
+    mockBindings = [
+      {
+        id: "bind1",
+        keys: ["KEY_LEFTMETA", "KEY_SPACE"],
+        gesture: "hold",
+        target_id: "default",
+        target_ids: ["default"],
+        tap_ms: 300,
+        hold_threshold_ms: 1000,
+        label: "Binding 1",
+        disabled: false,
+      },
+    ];
+
+    const { container } = render(HotkeysTab);
+
+    // Click Edit button to open Hotkey Editor modal
+    const editBtn = await screen.findByRole("button", { name: /Edit/i });
+    await fireEvent.click(editBtn);
+
+    // Verify Binding Editor modal is open
+    expect(screen.getByText("Edit Hotkey Binding")).not.toBeNull();
+
+    // Find the custom dropdown trigger button
+    const trigger = container.querySelector(".custom-select-trigger") as HTMLButtonElement;
+    expect(trigger).not.toBeNull();
+    expect(trigger.textContent).toContain("Focused Window");
+
+    // Click trigger to open dropdown list
+    await fireEvent.click(trigger);
+
+    // Click "-- Create New Target --" option button
+    const createBtn = screen.getByText("Create New Target");
+    expect(createBtn).not.toBeNull();
+    await fireEvent.click(createBtn);
+
+    // Verify Target Editor modal opens
+    expect(await screen.findByText("Create Target")).not.toBeNull();
+
+    // Verify that Target ID input is hidden in nested mode
+    const targetIdInput = screen.queryByPlaceholderText("e.g. obsidian_vault");
+    expect(targetIdInput).toBeNull();
+
+    // Click Cancel button in the Target modal
+    const cancelButtons = screen.getAllByRole("button", { name: /Cancel/i });
+    await fireEvent.click(cancelButtons[1]);
+
+    // Verify Target modal is closed
+    expect(screen.queryByText("Create Target")).toBeNull();
+
+    // Verify select visually reverts back to previous value
+    expect(trigger.textContent).toContain("Focused Window");
+  });
+
+  test("opens nested Target modal, creates a new target, and auto-selects it", async () => {
+    mockBindings = [
+      {
+        id: "bind1",
+        keys: ["KEY_LEFTMETA", "KEY_SPACE"],
+        gesture: "hold",
+        target_id: "default",
+        target_ids: ["default"],
+        tap_ms: 300,
+        hold_threshold_ms: 1000,
+        label: "Binding 1",
+        disabled: false,
+      },
+    ];
+
+    const { container } = render(HotkeysTab);
+
+    // Click Edit button to open Hotkey Editor modal
+    const editBtn = await screen.findByRole("button", { name: /Edit/i });
+    await fireEvent.click(editBtn);
+
+    // Find the custom dropdown trigger button
+    const trigger = container.querySelector(".custom-select-trigger") as HTMLButtonElement;
+    expect(trigger).not.toBeNull();
+
+    // Click trigger to open dropdown list
+    await fireEvent.click(trigger);
+    
+    // Click "-- Create New Target --" option button
+    const createBtn = screen.getByText("Create New Target");
+    await fireEvent.click(createBtn);
+
+    // Verify Target Editor modal opens
+    expect(await screen.findByText("Create Target")).not.toBeNull();
+
+    // Set Target display label
+    const labelInput = screen.getByPlaceholderText("e.g. Type directly into Obsidian");
+    await fireEvent.input(labelInput, { target: { value: "My Nested Target" } });
+
+    // Click Done to save target
+    const doneButtons = screen.getAllByRole("button", { name: /Done/i });
+    await fireEvent.click(doneButtons[1]);
+
+    // Verify Target modal is closed
+    expect(screen.queryByText("Create Target")).toBeNull();
+
+    // Verify select value was updated to the new target's label and delivery
+    expect(trigger.textContent).toContain("My Nested Target (inject)");
   });
 });
